refactor calling of newPassword dialogbox in shared folder
