+++ frontend/src/app/shared/components/newpassword/new-password.component.ts
