+++ frontend/src/app/shared/shared.module.ts
@@ -5,7 +5,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { HttpClientModule } from '@angular/common/http';
@@ -21,6 +21,7 @@ import { ErrorComponent } from './components/error/error.component';
 import { BackendService } from './services/backend.service';
 import { PageNavBarComponent } from './components/page-nav-bar/page-nav-bar.component';
 import { AutofocusDirective } from './directives/autofocus.directive';
+import { NewPasswordComponent } from './components/newpassword/new-password.component';
 
 @NgModule({
   imports: [
@@ -36,7 +37,8 @@ import { AutofocusDirective } from './directives/autofocus.directive';
     MatInputModule,
     HttpClientModule,
     MatTooltipModule,
-    MatButtonToggleModule
+    MatButtonToggleModule,
+    ReactiveFormsModule
   ],
   declarations: [
     ConfirmDialogComponent,
@@ -46,7 +48,8 @@ import { AutofocusDirective } from './directives/autofocus.directive';
     AlertComponent,
     ErrorComponent,
     PageNavBarComponent,
-    AutofocusDirective
+    AutofocusDirective,
+    NewPasswordComponent
   ],
   exports: [
     ConfirmDialogComponent,
@@ -56,7 +59,8 @@ import { AutofocusDirective } from './directives/autofocus.directive';
     AlertComponent,
     ErrorComponent,
     PageNavBarComponent,
-    AutofocusDirective
+    AutofocusDirective,
+    NewPasswordComponent
   ],
   providers: [
     BackendService
@@ -73,6 +77,7 @@ export { AlertComponent } from './components/alert/alert.component';
 export { CustomtextPipe } from './pipes/customtext/customtext.pipe';
 export { ConnectionStatus } from './interfaces/websocket-backend.interfaces';
 export { MainDataService } from './services/maindata/maindata.service';
+export { PasswordChangeService } from './services/password-change/password-change.service';
 export { BugReportService } from './services/bug-report.service';
 export { UserAgentService } from './services/user-agent/user-agent.service';
 export { SysConfig, AppSettings } from './interfaces/app-config.interfaces';
